refactor: 패키지 구조 변경
- 결합성이 높은 클래들을 같은 하위 패키지로 구성
