+++ src/main/java/ladder/domain/direction/Direction.java
@@ -1,0 +1,8 @@
+package ladder.domain.direction;
+
+public enum Direction {
+
+    LEFT,
+    NEUTRAL,
+    RIGHT
+}
