+++ src/main/java/ladder/domain/line/LineGenerator.java
@@ -1,0 +1,6 @@
+package ladder.domain.line;
+
+public interface LineGenerator {
+
+    public Line generateLine(int numberOfDirection);
+}
