--- src/main/java/ladder/domain/resource/direction/Direction.java
@@ -1,8 +1,0 @@
-package ladder.domain.resource.direction;
-
-public enum Direction {
-
-    LEFT,
-    NEUTRAL,
-    RIGHT
-}
