--- src/main/java/ladder/domain/resource/line/LineGenerator.java
@@ -1,6 +1,0 @@
-package ladder.domain.resource.line;
-
-public interface LineGenerator {
-
-    public Line generateLine(int numberOfDirection);
-}
