+++ src/test/java/ladder/domain/line/CustomLineGenerator.java
@@ -1,0 +1,16 @@
+package ladder.domain.line;
+
+public class CustomLineGenerator {
+
+//    @Override
+//    public Line generateLine() {
+//        return new Line();
+//    }
+//
+//    @Override
+//    public void insertDirectionIntoLine(Line line, int count) {
+//        line.addDirection(Direction.RIGHT);
+//        line.addDirection(Direction.LEFT);
+//        line.addDirection(Direction.NEUTRAL);
+//    }
+}
